2009-10-08  Ramon Novoa  <rnovoa@artica.es>
	* include/functions_ui.php,
          include/functions_incidents.php,
          operation/incidents/incident_detail.php: Security related fixes.
          Thanks to ~finalbeta and ~remove from the Global-Evolution Security
          Group.




git-svn-id: https://svn.code.sf.net/p/pandora/code/trunk@2012 c3f86ba8-e40f-0410-aaad-9ba5e7f4b01f

+++ pandora_console/operation/incidents/incident_detail.php
@@ -36,7 +36,7 @@ $actualizacion = get_system_time ();
 
 // EDITION MODE
 if (isset ($_GET["id"])) {
-	$id_inc = get_parameter_get ("id", 0);
+	$id_inc = (int) get_parameter_get ("id", 0);
 	
 	// Obtain group of this incident
 	$row = get_db_row ("tincidencia","id_incidencia",$id_inc);
